Use client date instead of sever date
of ffs javascript are you kidding me

Anyway, using the server date wasn't great because it meant we had to
deal with Time Zones. Using Javascript to determine the current date
should let us avoid that and provide the prompt for whatever today is
wherever the client is, which is what we actually want.