Load today's prompt without redirecting
I _really_ disliked that we redirected from the main index to show
today's prompt. Luckily, we can use JavaScript to fix this. When we hit
`/` now, we load the data from `/?date=<today's date>` and replace the
html with the result. This means we get to re-use our existing logic,
and layouts, and just replace the bit we care about. The end result is
that navigating to `/` shows today, no matter what, and we can still
pass the date as a param to show a specific day.